Move artworksSlice to artworks folder
