+++ src/artworks/artworksSlice/artworksSlice.ts
@@ -1,5 +1,4 @@
-import { createSlice } from "@reduxjs/toolkit";
-import type { PayloadAction } from "@reduxjs/toolkit";
+import { createSlice, PayloadAction } from "@reduxjs/toolkit";
 import { ArtworksState } from "./types";
 import { Artwork } from "../../artworks/types";
