+++ src/artworks/artworksSlice/types.ts
