+++ src/main.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import App from "./components/App/App";
 import { Provider } from "react-redux";
+import App from "./components/App/App";
 import { store } from "./redux/store";
 import "./styles/index.scss";
