+++ vite.config.ts
@@ -7,7 +7,7 @@ export default defineConfig({
   plugins: [react()],
   test: {
     globals: true,
-    setupFiles: ["./src/setupTest.ts"],
+    setupFiles: ["./src/setupTests.ts"],
     environment: "jsdom",
     coverage: {
       include: ["src/**/*.{ts,tsx}"],
